refactor: move patchbay degrade tests into module
